[MTS/LT][96823634] Fix thumbnail group transitions
https://www.pivotaltracker.com/story/show/96823634

When you used the main image gallery controls to transition from one
image to the next, the thumbnail groups would keep in sync with the
main images. This did not work when transitioning from the second
thumbnail group back to the first thumbnail group. The problem happened
only when the thumbnails were hidden. Swiper relied on WebKitCSSMatrix,
which wasn't returning the values we wanted when the thumbnails weren't
visibe. See
https://github.com/nolimits4web/Swiper/blob/v2.7.5/src/idangerous.swiper.js#L2575
if you're interested in Swiper's implementation.

Our solution is to show the drawer, let Swiper do its calculations, and
then hide the drawer. This happens fast enough that the user can't see
the drawer being shown, and then hidden again.