[HH/LT][103664656] Remove uiGallerySlideClicked Event
Using the uiGallerySlideClicked event to sync the thumbnails with a main
Gallery is causing all galleries on the page to attempt to sync as well.

This removes the event from ui/thumbnails and ui/gallery and changes
ui/thumbnails to trigger a uiGallerySlideChanged event so that a
ui/gallery_syncer can be used instead to sync a specific set of
Galleries.